+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 - Generate random password
 - Print the expiration time of TOTP code
 - Export a wallet in json
+- Import entries from a json file
 
 ## Changed